+++ inject/inject-thrift-client/src/main/scala/com/twitter/inject/thrift/ThriftMethodBuilder.scala
@@ -1,10 +1,10 @@
 package com.twitter.inject.thrift
 
+import com.twitter.conversions.StringOps._
 import com.twitter.finagle.service.ResponseClassifier
 import com.twitter.finagle.thrift.service.{Filterable, ReqRepServicePerEndpointBuilder, ServicePerEndpointBuilder}
 import com.twitter.finagle.thriftmux.MethodBuilder
 import com.twitter.finagle.{Filter, Service}
-import com.twitter.inject.conversions.string._
 import com.twitter.inject.Injector
 import com.twitter.inject.thrift.internal.filters.ThriftClientExceptionFilter
 import com.twitter.scrooge.ThriftMethod
@@ -266,7 +266,7 @@ final class ThriftMethodBuilder[ServicePerEndpoint <: Filterable[ServicePerEndpo
       implementation
         .getClass
         .getDeclaredMethods
-        .find(_.getName == method.name.camelify) // converts a snake_case ThriftMethod.name to camelCase for reflection method lookup
+        .find(_.getName == method.name.toCamelCase) // converts a snake_case ThriftMethod.name to camelCase for reflection method lookup
     methodOpt match {
       case Some(serviceMethod) =>
         serviceMethod