+++ inject/inject-utils/src/main/scala/com/twitter/inject/conversions/string.scala
@@ -1,27 +1,27 @@
 package com.twitter.inject.conversions
 
-import com.twitter.inject.utils.StringUtils
-import org.apache.commons.lang.{StringUtils => ApacheCommonsStringUtils}
+import com.twitter.conversions.StringOps
+import org.apache.commons.{lang => acl}
 
 object string {
 
   implicit class RichString(val self: String) extends AnyVal {
-    def toOption = {
+    def toOption: Option[String] = {
       if (self == null || self.isEmpty)
         None
       else
         Some(self)
     }
 
-    def getOrElse(default : => String) = {
+    def getOrElse(default : => String): String = {
       if (self == null || self.isEmpty)
         default
       else
         self
     }
 
-    def ellipse(len: Int) = {
-      ApacheCommonsStringUtils.abbreviate(self, len + 3) // adding 3 for the ellipses :-/
+    def ellipse(len: Int): String = {
+      acl.StringUtils.abbreviate(self, len + 3) // adding 3 for the ellipses :-/
     }
 
     /**
@@ -35,9 +35,8 @@ object string {
      *
      * will return the String `"helloWorld"`.
      */
-    def camelify = {
-      StringUtils.camelify(self)
-    }
+    @deprecated("Users are encouraged to use com.twitter.conversions.StringOps#toCamelCase", "2019-03-02")
+    def camelify: String = StringOps.toCamelCase(self)
 
     /**
      * Converts foo_bar to FooBar (first letter uppercased)
@@ -50,9 +49,8 @@ object string {
      *
      * will return the String `"HelloWorld"`.
      */
-    def pascalify = {
-      StringUtils.pascalify(self)
-    }
+    @deprecated("Users are encouraged to use com.twitter.conversions.StringOps#toPascalCase", "2019-03-02")
+    def pascalify: String = StringOps.toPascalCase(self)
 
     /**
      * Converts FooBar to foo_bar (all lowercase)
@@ -65,8 +63,7 @@ object string {
      *
      * will return the String `"hello_world"`.
      */
-    def snakify = {
-      StringUtils.snakify(self)
-    }
+    @deprecated("Users are encouraged to use com.twitter.conversions.StringOps#toSnakeCase", "2019-03-02")
+    def snakify: String = StringOps.toSnakeCase(self)
   }
 }